fix: move ABS property adding to it's own class
